adds functionality to add get all cards by deckid as callback, started implement new deck screen

+++ Model/DeckModel.swift
@@ -8,6 +8,6 @@
 import Foundation
 
 struct DeckModel {
-    var id: Int64
+    var id: Int
     var deckName : String
 }
